# lib/api-zod/src/generated/types/jobApplicationList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobApplication } from "./jobApplication";

export interface JobApplicationList {
  applications: JobApplication[];
}