# lib/api-zod/src/generated/types/pipelineStageCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobApplicationStatus } from "./jobApplicationStatus";

export interface PipelineStageCount {
  status: JobApplicationStatus;
  /** @minimum 0 */
  count: number;
}