# lib/api-zod/src/generated/types/pipelineSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PipelineRecentActivity } from "./pipelineRecentActivity";
import type { PipelineStageCount } from "./pipelineStageCount";

export interface PipelineSummary {
  /** @minimum 0 */
  total: number;
  stages: PipelineStageCount[];
  recentActivity: PipelineRecentActivity[];
}