# lib/api-zod/src/generated/types/updateJobApplicationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobApplicationStatus } from "./jobApplicationStatus";

export interface UpdateJobApplicationRequest {
  /** @minLength 1 */
  company?: string;
  /** @minLength 1 */
  role?: string;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  salaryMin?: string | null;
  /** @nullable */
  salaryMax?: string | null;
  /** @nullable */
  source?: string | null;
  status?: JobApplicationStatus;
}